EBDA corrections by Lucho
git-svn-id: https://svn.code.sf.net/p/freedos/svn/kernel/trunk@680 6ac86273-5f31-0410-b378-82cca8765d1b

+++ kernel/init-mod.h
@@ -142,7 +142,7 @@ VOID FAR * HMAalloc(COUNT bytesToAllocate);
 
 /* initoem.c */
 unsigned init_oem(void);
-void movebda(int bytes, unsigned new_seg);
+void movebda(size_t bytes, unsigned new_seg);
 unsigned ebdasize(void);
 
 /* intr.asm */
